feat: Config file the logging module to use Pydantic
